refactor: rename Connect component to Collaborate and update import paths
